+++ src/bpp/management/commands/rebuild_cache.py
@@ -16,7 +16,11 @@ from bpp.models import (
     Praca_Habilitacyjna,
     Praca_Doktorska,
 )
-from bpp.util import no_threads, partition_count
+from bpp.util import (
+    no_threads,
+    partition_count,
+    disable_multithreading_by_monkeypatching_pool,
+)
 
 
 def subprocess_setup(*args):
@@ -28,7 +32,11 @@ def subprocess_setup(*args):
 class Command(BaseCommand):
     help = "Odbudowuje cache"
 
-    def handle(self, *args, **options):
+    def add_arguments(self, parser):
+        parser.add_argument("--disable-multithreading", action="store_true")
+
+    def handle(self, disable_multithreading, *args, **options):
+
         if not settings.TESTING:
             from django import db
 
@@ -36,6 +44,8 @@ class Command(BaseCommand):
 
         pool_size = no_threads(0.75)
         pool = multiprocessing.Pool(processes=pool_size, initializer=subprocess_setup)
+        if disable_multithreading:
+            disable_multithreading_by_monkeypatching_pool(pool)
 
         pc = pool.apply(partition_count, args=(Praca_Habilitacyjna.objects, pool_size))
         pool.starmap(rebuild_praca_habilitacyjna, pc)