+++ src/bpp/tests/test_ranking_autorow.py
@@ -1,0 +1,27 @@
+from bpp.views.raporty import RankingAutorow
+
+
+def test_ranking_autorow_get_queryset_prace_sa(
+    wydawnictwo_ciagle_z_autorem,
+    wydawnictwo_zwarte_z_autorem,
+    rf,
+    uczelnia,
+    przed_korekta,
+):
+    wydawnictwo_ciagle_z_autorem.punkty_pk = 20
+    wydawnictwo_ciagle_z_autorem.impact_factor = 20
+    wydawnictwo_ciagle_z_autorem.status_korekty = przed_korekta
+    wydawnictwo_ciagle_z_autorem.save()
+
+    wydawnictwo_zwarte_z_autorem.impact_factor = 50
+    wydawnictwo_zwarte_z_autorem.punkty_pk = 20
+    wydawnictwo_zwarte_z_autorem.status_korekty = przed_korekta
+    wydawnictwo_ciagle_z_autorem.save()
+
+    r = RankingAutorow(request=rf.get("/"), kwargs=dict(od_roku=0, do_roku=3030))
+
+    assert r.get_queryset().count() == 1
+
+    uczelnia.ukryj_status_korekty_set.create(status_korekty=przed_korekta)
+
+    assert r.get_queryset().count() == 0